Move stream encoder tests to their own file
