+++ src/write/mod.rs
@@ -1,3 +1,6 @@
 //! Implementations of `io::Write` to transparently handle base64 encoding.
 mod encoder;
 pub use self::encoder::Base64Encoder;
+
+#[cfg(test)]
+mod encoder_tests;
